refactor: Delimiter 객체 생성
input에서 delimiter를 찾는 과정을 객체로 만들고 캡슐화했습니다.
