+++ src/main/java/calculator/InputSequence.java
@@ -1,33 +1,20 @@
 package calculator;
 
-import java.math.BigInteger;
 import java.util.Arrays;
 import java.util.List;
 
 public class InputSequence {
     public static final long SUM_INITIAL_VALUE = 0L;
-    public static final int DELIMITER_START_INDEX = 2;
     private final List<Long> sequence;
 
     public InputSequence(String input) {
-        String delimiter = validateCustomDelimiter(input);
-        this.sequence = Arrays.stream(input.split(delimiter))
+        Delimiter delimiter = new Delimiter(input);
+        this.sequence = Arrays.stream(input.split(delimiter.get()))
                             .map(part -> {
                                 validateSplit(part);
                                 return castingElementToLong(part);
                             })
                             .toList();
-    }
-
-    private String validateCustomDelimiter(String input) {
-        if (input.startsWith("//")) {
-            for (int i = DELIMITER_START_INDEX; i < input.length(); i++) {
-                if (input.startsWith("\\n", i)) {
-                    return input.substring(DELIMITER_START_INDEX, i);
-                }
-            }
-        }
-        return "[,:]";
     }
 
     private void validateSplit(String input) {
